pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "markdf"
version = "0.1.0"
description = "把 Markdown 渲染成排版精致的 A4 PDF（可主题扩展）"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "markdf" }]
dependencies = [
  "jinja2>=3.1.3",
  "linkify-it-py>=2.0.3",
  "markdown-it-py>=3.0.0",
  "mdit-py-plugins>=0.4.2",
  "playwright>=1.41.2",
  "pygments>=2.17.2",
]

[project.scripts]
markdf = "markdf.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/markdf"]
include = ["src/markdf/static/**", "src/markdf/themes/**"]

[tool.pytest.ini_options]
testpaths = ["tests"]
